% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_calculator.R
\name{CI_calculator}
\alias{CI_calculator}
\title{Confidence Interval calculator}
\usage{
CI_calculator(exp_data, experiment_data, p)
}
\arguments{
\item{exp_data}{Input data}

\item{experiment_data}{experiment data}

\item{p}{A float between 0 and 1 indicating the percentage used for constructing the CI}
}
\value{
The CI for the probabilities
}
\description{
Calculate the CI for the probabilities
}
