% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preperation.R
\name{data_preparation}
\alias{data_preparation}
\title{Data Preparation
Normalizes the PT}
\usage{
data_preparation(data)
}
\arguments{
\item{data}{the input data file which is a data frame}
}
\value{
A data frame with the normalized PT
}
\description{
Data Preparation
Normalizes the PT
}
