% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_comparison.R
\name{gen_comparison}
\alias{gen_comparison}
\title{Probability Extractor}
\usage{
gen_comparison(Dataset, Experiment, Ranks_table)
}
\arguments{
\item{Dataset}{Input data}

\item{Experiment}{The experiment}

\item{Ranks_table}{table of ranks}
}
\value{
a list of 2 items:
\itemize{
   \item df1
   \item Rey
}
}
\description{
Finds the probability of ranks for each genotype in all bootstrap resamples.
}
