% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_list_probs.R
\name{genotype_list_probs}
\alias{genotype_list_probs}
\title{Distribution of rank probabilities}
\usage{
genotype_list_probs(subset_exp, boots_train)
}
\arguments{
\item{subset_exp}{The input file}

\item{boots_train}{training bootstrap resamples}
}
\value{
probabilities for different varieties
}
\description{
Calculates how the rank probabilities are distributed
}
