% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_probs.R
\name{pairwise_probs}
\alias{pairwise_probs}
\title{Pairwise Comparison}
\usage{
pairwise_probs(subset_exp, Rank_Table, index_file, genotype_set)
}
\arguments{
\item{subset_exp}{Input data}

\item{Rank_Table}{table of ranks}

\item{index_file}{matrix of the bootstrap resample of location}

\item{genotype_set}{vector of genotypes of interest}
}
\value{
pairwise comparison
}
\description{
Creates pairwise probability comparison
}
