% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_f.R
\name{prob_f}
\alias{prob_f}
\title{probaility f}
\usage{
prob_f(subset_exp, new_analysis, u, Nonchecks = FALSE)
}
\arguments{
\item{subset_exp}{inut data}

\item{new_analysis}{new analysis}

\item{u}{genotype to compare}

\item{Nonchecks}{Nonchecks}
}
\value{
rank & probs
}
\description{
helper function
}
