% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilistic_Ranking.R
\name{Probabilistic_Ranking}
\alias{Probabilistic_Ranking}
\title{Probabilistic Ranking Function}
\usage{
Probabilistic_Ranking(
  experiment_data,
  output_name,
  genotype_set = NULL,
  boots_matrix = NULL
)
}
\arguments{
\item{experiment_data}{input data}

\item{output_name}{Name of the file to be stored in /Outputs/ directory.}

\item{genotype_set}{vector of genotypes of interest}

\item{boots_matrix}{matrix of bootstrap resamples of locations (default = NULL).
By default it generates a bootstrap resample with 1000 iteration.}
}
\value{
A list of 2 items:
\itemize{
    \item pairwise_probabilities
    \item probabilistic_ranks
}
}
\description{
Doing some analysis on the ranks and generating the desired output using the probabilistic method
}
