% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probabilistic_ranks.R
\name{Probabilistic_ranks}
\alias{Probabilistic_ranks}
\title{Rank Analyzer Function}
\usage{
Probabilistic_ranks(subset_exp, pairwise_probabilities)
}
\arguments{
\item{subset_exp}{input data}

\item{pairwise_probabilities}{pairwise comparison probabilities}
}
\value{
A table of probabilistic ranks
}
\description{
Doing some analysis on the ranks and generating the desired output
}
