% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_analyzer.R
\name{rank_analyzer}
\alias{rank_analyzer}
\title{Rank Analyzer Function}
\usage{
rank_analyzer(
  experiment_data,
  output_name,
  genotype_set = "",
  boots_matrix = NULL,
  method = "both"
)
}
\arguments{
\item{experiment_data}{input data}

\item{output_name}{Name of the file to be stored in /Outputs/ directory.}

\item{genotype_set}{vector of genotypes of interest}

\item{boots_matrix}{matrix of bootstrap resamples of locations (default=NULL).
By default it generates a bootstrap resample with 1000 iteration.}

\item{method}{the method to do the analysis based on it}
}
\value{
A list of 6 items:
\itemize{
\item CI_data
\item CI_Info
\item boost_df_obs
\item boost_PT_info_obs
\item pairwise_probabilities
\item probabilistic_ranks
}
}
\description{
Doing some analysis on the ranks and generating the desired output
}
