% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank_CI.R
\name{Rank_CI}
\alias{Rank_CI}
\title{Rank CI Function}
\usage{
Rank_CI(
  experiment_data,
  output_name = NULL,
  genotype_set = NULL,
  boots_matrix = NULL,
  p = NULL
)
}
\arguments{
\item{experiment_data}{input data}

\item{output_name}{Name of the file to be stored in /Outputs/ directory.}

\item{genotype_set}{vector of genotypes of interest}

\item{boots_matrix}{matrix of bootstrap resamples of locations (default = NULL).
By default it generates a bootstrap resample with 800 iteration.}

\item{p}{confidence interval (default = 0.8)}
}
\value{
A list of 6 items:
\itemize{
    \item rank_probabilities
    \item rank_info
    \item bs_ranks
    \item bs_PTs
    \item RI
    \item rankCI
}
}
\description{
Doing some analysis on the ranks and generating the desired output using mean phenotype method
}
