% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_visualization.R
\name{rank_visualization}
\alias{rank_visualization}
\title{Rank Visualizer}
\usage{
rank_visualization(input_name, p, n_top = 2, method = "mean_phenotype")
}
\arguments{
\item{input_name}{Name of the saved file from \code{rank_analyzer} function which is stored}

\item{p}{A float in \link{0,1} used for constructing the confidence interval}

\item{n_top}{An integer value showing the number of top genotypes to be displayed (default=2)
in /Outputs/ directory.}
}
\value{
a ggplot object of the heatmap for the CI of the ranks.
}
\description{
Plots the heatmap of the CI's for probabilities
}
