% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPhenotye_extractor.R
\name{RankPhenotype_extractor}
\alias{RankPhenotype_extractor}
\title{Rank & Phenotype Extractor}
\usage{
RankPhenotype_extractor(subset_exp, include_env)
}
\arguments{
\item{subset_exp}{input data}

\item{include_env}{include environment or not}
}
\value{
Rank table & Yield table
}
\description{
Extracts the rank and yield tables
}
