% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_correction.R
\name{rm_correction}
\alias{rm_correction}
\title{RM Correction}
\usage{
rm_correction(data)
}
\arguments{
\item{data}{the preprocessed file from the output of function}
}
\value{
Returns the data with a new feature that describes the corrected PT
}
\description{
It removes the effect of the locations
}
