% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_oat.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample dataset for the genotype ranking}
\format{
A dataframe with 1688 observations and 10 variables:
\describe{
\item{X}{X}
\item{VARIETY}{variety of genotypes}
\item{LOCATION}{location}
\item{group}{group}
\item{PT}{amount of phenotype}
\item{env_index}{environment index}
\item{MINTF}{maize race}
\item{RADG}{radg}
\item{YEAR}{year}
\item{CHECK}{whether the genotype is one of the baselines or not}
}

A dataframe with 1688 observations and 10 variables:
\describe{
\item{X}{X}
\item{GENOTYPE}{genotypes planted in a trial}
\item{ENVIRONMENT}{location-year the crop genotype is planted in}
\item{REPNO}{replications of each genotype in each environment}
\item{PT}{phenotypic of phenotype}
\item{YEAR}{year}
\item{CHECK}{whether the genotype is one of the baselines or not}
\item{LOCATION}{location the crop genotype is planted in}
}
}
\usage{
sample_data

sample_data
}
\description{
A dataset containing the genotype and the locations and phenotype..

A dataset containing the genotype and the locations and phenotype.
}
\keyword{datasets}
