% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_function_withcheck.R
\name{win_function_withcheck}
\alias{win_function_withcheck}
\title{Win function with check}
\usage{
win_function_withcheck(Rank_Table_train1, var1)
}
\arguments{
\item{Rank_Table_train1}{Input data}

\item{var1}{variable 1}
}
\value{
pdata
}
\description{
do the same as gene_comparison but with checks
}
